pyproject.toml
[tool.poetry]
name = "vk-to-tgm"
version = "1.0.0"
description = "An application that forwards wall posts and playlists from VK community to Telegram channel"
license = "GPL-3.0-or-later"
authors = ["Ilia Boyazitov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.2"
vkbottle = "^4.2.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
ffmpeg-python = "^0.2.0"
gunicorn = "^20.1.0"
celery = "^5.2.6"
SQLAlchemy = "^1.4.35"
Telethon = "^1.24.0"
cryptg = "^0.2.post4"
Pillow = "^9.1.0"
aiohttp = "^3.8.1"
hachoir = "^3.1.2"
eyed3 = "^0.9.6"
aiofiles = "^0.8.0"
vkaudiotoken = {git = "https://github.com/Bizordec/vkaudiotoken-python.git"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
flake8-bugbear = "^22.4.25"
flake8-comprehensions = "^3.8.0"
flake8-tidy-imports = "^4.6.0"
sqlalchemy2-stubs = "^0.0.2-alpha.22"
bandit = "^1.7.4"
Babel = "^2.10.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
